+++ utility/test/module_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/golang/mock/gomock"
 	"github.com/pokt-network/pocket/persistence"
+	"github.com/pokt-network/pocket/runtime/defaults"
 	"github.com/pokt-network/pocket/runtime/test_artifacts"
 	"github.com/pokt-network/pocket/shared/modules"
 	mock_modules "github.com/pokt-network/pocket/shared/modules/mocks"
@@ -27,7 +28,7 @@ var (
 	defaultTestingChainsEdited = []string{"0002"}
 	defaultUnstaking           = int64(2017)
 	defaultSendAmount          = big.NewInt(10000)
-	defaultNonceString         = utilTypes.BigIntToString(test_artifacts.DefaultAccountAmount)
+	defaultNonceString         = utilTypes.BigIntToString(defaults.DefaultAccountAmount)
 	defaultSendAmountString    = utilTypes.BigIntToString(defaultSendAmount)
 	testSchema                 = "test_schema"
 	testMessageSendType        = "MessageSend"
@@ -81,6 +82,7 @@ func newTestPersistenceModule(t *testing.T, databaseUrl string) modules.Persiste
 	mockPersistenceConfig.EXPECT().GetPostgresUrl().Return(databaseUrl).AnyTimes()
 	mockPersistenceConfig.EXPECT().GetNodeSchema().Return(testSchema).AnyTimes()
 	mockPersistenceConfig.EXPECT().GetBlockStorePath().Return("").AnyTimes()
+	mockPersistenceConfig.EXPECT().GetTxIndexerPath().Return("").AnyTimes()
 
 	mockRuntimeConfig := mock_modules.NewMockConfig(ctrl)
 	mockRuntimeConfig.EXPECT().GetPersistenceConfig().Return(mockPersistenceConfig).AnyTimes()