+++ utility/doc/CHANGELOG.md
@@ -7,7 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.7] - 2022-10-12
+## [0.0.0.9] - 2022-11-01
+
+- Remove `TxResult` from the utility module interface (added in TxIndexer integration of transaction indexer (issue-#168) #302)
+- Combined creation and application of block in proposer lifecycle
+
+## [0.0.0.8] - 2022-10-17
+
+- Added Relay Protocol interfaces and diagrams
+
+## [0.0.0.7] - 2022-10-14
+
+- Added session interfaces and diagrams
 - Moved `TxIndexer` package to persistence module
 - Added new proto structure `DefaultTxResult`
 - Integrated the `TxIndexer` into the lifecycle